Fix clang warnings for linux builds (#6140)
https://github.com/youtube/cobalt/pull/6139 will actually remove the
clang warning suppression flags after this is submitted.

Bug: 374302273
Bug: 378934120

original-hexsha: 274beb9dd42dc6c7d7dd3ff9415a41e1393f0133

+++ starboard/nplb/posix_compliance/posix_file_get_path_info_test.cc
@@ -91,7 +91,7 @@ TEST(PosixFileGetPathInfoTest, WorksOnStaticContentFiles) {
     struct stat info;
     EXPECT_TRUE(stat(filename.c_str(), &info) == 0);
     size_t content_length = GetTestFileExpectedContent(filename).length();
-    EXPECT_EQ(content_length, info.st_size);
+    EXPECT_EQ(static_cast<long>(content_length), info.st_size);
     EXPECT_FALSE(S_ISDIR(info.st_mode));
     EXPECT_FALSE(S_ISLNK(info.st_mode));
   }
